+++ R/name_from_yaml.R
@@ -1,5 +1,11 @@
 name_from_yaml <- function(yaml_path,
                            verbose = TRUE) {
+    if(!file.exists(yaml_path)){
+        # messager(yaml_path,"does not exist.",
+        #          "Inferring conda_env name from path.",v=verbose)
+        conda_env <- gsub(".yaml$|.yml$","",basename(yaml_path))
+        return(conda_env)
+    }
     messager("echoconda:: Retrieving conda env name from yaml.", v = verbose)
     lines <- suppressWarnings(tryCatch(
         expr = {